refactor out script tools
